caprevoke: Avoid validating CLGs if revocation is in progress
When forking, the GCLG and LCLGs may be out of sync if a new revocation epoch
was opened (bumping the GCLG) but the background scan has not yet started
(background scans block fork()).  This can arise if a cheri_revoke() is called
without either of the CHERI_REVOKE_LAST_PASS and CHERI_REVOKE_ASYNC flags.