kasan: Temporarily change KASAN intercepted {memcpy,memset,memmove} to macros
This requires collaboration with the compiler in that the
compiler ASan intrumentation pass inserts a function call
to the interceptor function but doesn't erase the original
function call. The rationale behind this commit is that in
Morello LLVM __builtin_{memset,memcpy,memmove} have
different subobject-bounds semantics to a function wrapper
to these builtins. The bounds are assumed to be of the
container, as in:
```c
	bcopy(&src->cr_startcopy, &dest->cr_startcopy,
	    (unsigned)((caddr_t)&src->cr_endcopy -
		(caddr_t)&src->cr_startcopy));
```
cr_startcopy to cr_endcopy span many object fields, but
a call to builtin memcpy will not trigger subobject bounds
SIGPROT. But a call to a wrapper will crash due to
cr_startcopy having the bounds of the first field.
```c
void *wrapper_memcpy(void *dst, void *src, size_t len)
{  return __builtin_memcpy(dst, src, len); }
```
A future commit could mark such code patterns in the
kernel to relax subobject bounds checking.