kasan: Add an initial working version of KASAN quarantining
It is tested on Morello only and only using kernel boot. More
testing required. This feature delays the reuse of UMA
allocations by quarantining them.

The quarantining is based on hooking uma_zfree_arg calls
by saving the arguments to uma_zfree_arg after running
the item destructor, which poisons the UMA allocation to free.
The arguments are queued to a per-CPU quarantine and
they are exchanged with the uma_zfree_arg invocation when
the quarantine is "full" by reaching a maximum number of
entries. Each quarantined item incurs the same amount of
space corresponding to sizeof(struct
kasan_quarantine_item). The quarantine can be configured
to have a different max number of entries by using a macro.

This feature does not yet support quarantining SMR zones.