kasan: Implement quarantining for SMR zones
The current implementation uses a separate PCPU quarantine
shared for all SMR zones. For SMR zones,
kasan_quarantine_put() is invoked when:
1. The item is freed directly to the zone.
2. A bucket is drained. All the items in the bucket are
quarantined, and popped items are individually freed.
3. A bucket is retrieved from one of the zone's caches. In this
case, we free all cached items and free the bucket, instead of
reusing the cache bucket. This simplifies the quarantine
design but it is not ideal for performance.

KASAN quarantining of SMR zones can be optionally turned
on by defining the kernel option KASAN_QUARANTINE_SMR,
potentially incurring a non-negligible performance cost.